+++ views/data/livestats.go
@@ -1,4 +1,4 @@
-package viewdata
+package data
 
 import (
 	"errors"
